+++ src/App.jsx
@@ -1,12 +1,11 @@
-// App.js
 import React from "react";
-import Home from "./Home";
-import Profile from './Profile'
-import FrenchProfile from "./FrenchProfile";
-import FormLayout from './FormLayout'
+import Home from "./Pages/Home.jsx";
+import Profile from './Pages/Profile.jsx'
+import FrenchProfile from "./Pages/FrenchProfile.jsx";
+import FormLayout from './Pages/FormLayout.jsx'
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Header from './Header';
-import Footer from "./Footer";
+import Header from './Components/Header';
+import Footer from "./Components/Footer";
 
 function App() {
   return (
