--- src/Next.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export const Next = () => {
-  return (
-    <button>Next</button>
-  )
-}
